# src/main/resources/api-docs.yaml
openapi: 3.0.0
info:
  title: Moodi-pie API
  description: 스웨거 API
  version: v1.0.0
servers:
  - url: https://moodipie.shop
  - url: https://운영URL
...
paths:
...
  /playlists/{playlistNumber}:
    get:
      tags:
        - 플레이리스트 CRUD
      summary: 플레이리스트 정보 조회
      description: 생성된 플레이리스트를 조회합니다.
      operationId: playlistResponse
      parameters:
        - name: playlistNumber
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 조회 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResPlaylistTrackResponse"
        "400":
          description: 요청 실패
          content:
            application/json:
              examples:
                조회 실패 - playlistNumber NULL:
                  summary: ""
                  description: 잘못된 playlistNumber로 조회하여 플레이리스트 조회에 실패했습니다.
                  value:
                    httpStatus: 400
                    message: NULL 값이 존재합니다.
                    divisionCode: G003
      security:
        - Authorization: []
    put:
      tags:
        - 플레이리스트 CRUD
      summary: 플레이리스트 정보 수정
      description: 생성된 플레이리스트를 수정합니다.
      operationId: updatePlaylist
      parameters:
        - name: playlistNumber
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlaylistRequest"
        required: true
      responses:
        "200":
          description: 수정 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResUpdatePlaylistRequest"
        "400":
          description: 요청 실패
          content:
            application/json:
              examples:
                수정 실패 - playlistNumber NULL:
                  summary: ""
                  description: 잘못된 playlistNumber로 수정을 시도하여 플레이리스트 수정에 실패했습니다.
                  value:
                    httpStatus: 400
                    message: NULL 값이 존재합니다.
                    divisionCode: G003
                수정 실패 - 필드 에러:
                  summary: ""
                  description: 필드 조건에 맞지 않아 플레이리스트 수정에 실패했습니다.
                  value:
                    httpStatus: 400
                    message: 잘못된 형식입니다.
                    divisionCode: G002
      security:
        - Authorization: []
    delete:
      tags:
        - 플레이리스트 CRUD
      summary: 플레이리스트 삭제
      description: 홈 화면에서 플레이리스트를 삭제합니다.
      operationId: deletePlaylist
      parameters:
        - name: playlistNumber
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 삭제 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResPlaylistResponse"
        "400":
          description: 요청 실패
          content:
            application/json:
              examples:
                삭제 실패 - playlistNumber NULL:
                  summary: ""
                  description: 잘못된 playlistNumber 로 삭제를 시도하여 플레이리스트 삭제에 실패했습니다.
                  value:
                    httpStatus: 400
                    message: NULL 값이 존재합니다.
                    divisionCode: G003
      security:
        - Authorization: []
...
  /playlists:
    get:
      tags:
        - 플레이리스트 CRUD
      summary: 홈에서 플레이리스트 조회
      description: 홈 화면에서 플레이리스트를 조회합니다.
      operationId: findAllPlaylist
      responses:
        "200":
          description: 조회 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResListPlaylistResponse"
        "400":
          description: 요청 실패
          content:
            application/json:
              examples:
                조회 실패 - userEmail 오류:
                  summary: ""
                  description: userEmail 값이 Null 이라서 플레이리스트 조회에 실패했습니다.
                  value:
                    httpStatus: 400
                    message: NULL 값이 존재합니다.
                    divisionCode: G003
      security:
        - Authorization: []
    post:
      tags:
        - 플레이리스트 CRUD
      summary: 플레이리스트 생성
      description: 새로운 플레이리스트를 생성합니다.
      operationId: savePlaylist
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlaylistRequest"
        required: true
      responses:
        "201":
          description: 생성 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResPlaylistResponse"
        "400":
          description: 요청 실패
          content:
            application/json:
              examples:
                생성 실패 - 플레이리스트 NULL:
                  summary: ""
                  description: 요청값이 없어 플레이리스트가 생성되지 않았습니다.
                  value:
                    httpStatus: 400
                    message: 잘못된 형식입니다.
                    divisionCode: G002
        "404":
          description: 리소스를 찾을 수 없음
          content:
            application/json:
              examples:
                생성 실패 - 스포티파이 오류:
                  summary: ""
                  description: 스포티파이 오류로 플레이리스트가 생성되지 않았습니다.
                  value:
                    httpStatus: 404
                    message: 스포티파이에서 곡을 찾지 못했습니다.
                    divisionCode: S001
      security:
        - Authorization: []
...
  /playlists/{playlistNumber}/tracks/{playlistTrackNumber}:
    delete:
      tags:
        - 플레이리스트 CRUD
      summary: 플레이리스트 내부 트랙 삭제
      description: 플레이리스트 내부의 트랙을 삭제 합니다.
      operationId: deletePlaylistTrack
      parameters:
        - name: playlistNumber
          in: path
          description: 선택할 플레이리스트 아이디
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: playlistTrackNumber
          in: path
          description: 삭제할 트랙 번호
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 삭제 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResPlaylistTrackEachResponse"
        "400":
          description: 요청 실패
          content:
            application/json:
              examples:
                삭제 실패 - playlistNumber or trackId NULL:
                  summary: ""
                  description: 잘못된 playlistNumber 나 TrackId로 삭제를 시도하여 플레이리스트 삭제에 실패했습니다.
                  value:
                    httpStatus: 400
                    message: NULL 값이 존재합니다.
                    divisionCode: G003
      security:
        - Authorization: []
components:
  schemas:
...
    FieldErrors:
      type: object
      properties:
        field:
          type: string
        value:
          type: string
        reason:
          type: string
...
    UpdatePlaylistRequest:
      required:
        - emotion
        - playlistImage
        - title
      type: object
      properties:
        title:
          type: string
          description: 플레이리스트 제목
          example: 우울할때 듣는 음악
        playlistImage:
          type: string
          description: 플레이리스트 이미지 URL
          example: http://dfdfjiweoif.com
        emotion:
          type: string
          description: 감정
          example: SAD
          enum:
            - HAPPY
            - SAD
            - MAD
            - EXCITING
            - TIRED
    ApiResUpdatePlaylistRequest:
      title: API 응답
      type: object
      properties:
        httpStatus:
          type: integer
          description: 응답코드
          format: int32
        data:
          $ref: "#/components/schemas/UpdatePlaylistRequest"
        message:
          type: string
          description: 응답 메세지
          example: ex.조회에 성공했습니다.
        divisionCode:
          type: string
          description: 오류 구분 코드
          example: 오류시 출력
        errors:
          type: array
          description: 상세 필드 에러
          example: 오류시 출력
          items:
            $ref: "#/components/schemas/FieldErrors"
...
    CreatePlaylistRequest:
      required:
        - emotion
        - playlistImage
        - title
        - trackIds
      type: object
      properties:
        title:
          type: string
          description: 플레이리스트 제목
          example: 우울할때 듣는 음악
        playlistImage:
          type: string
          description: 플레이리스트 이미지 URL
          example: http://dfdfjiweoif.com
        emotion:
          type: string
          description: 감정
          example: SAD
          enum:
            - HAPPY
            - SAD
            - MAD
            - EXCITING
            - TIRED
        trackIds:
          type: array
          description: 플레이리스트 트랙 ID들
          example:
            - 7ouMYWpwJ422jRcDASZB7P
            - 4VqPOruhp5EdPBeR92t6lQ
            - 2takcwOaAZWiXQijPHIx7B
          items:
            type: string
    ApiResPlaylistResponse:
      title: API 응답
      type: object
      properties:
        httpStatus:
          type: integer
          description: 응답코드
          format: int32
        data:
          $ref: "#/components/schemas/PlaylistResponse"
        message:
          type: string
          description: 응답 메세지
          example: ex.조회에 성공했습니다.
        divisionCode:
          type: string
          description: 오류 구분 코드
          example: 오류시 출력
        errors:
          type: array
          description: 상세 필드 에러
          example: 오류시 출력
          items:
            $ref: "#/components/schemas/FieldErrors"
    PlaylistResponse:
      type: object
      properties:
        playlistNumber:
          type: integer
          description: 플레이 리스트 넘버
          format: int64
          example: 1
        title:
          type: string
          description: 제목
          example: http://dfdfjiweoif.com
        playlistImage:
          type: string
          description: 플레이리스트 이미지 URL
          example: http://dfdfjiweoif.com
        timestamp:
          type: string
          description: 플레이리스트 생성 시간
          format: date
          example: 2024-12-12
        emotion:
          type: string
          description: 감정
          example: SAD
          enum:
            - HAPPY
            - SAD
            - MAD
            - EXCITING
            - TIRED
...
    TrackResponse:
      type: object
      properties:
        playlistTrackNumber:
          type: integer
          format: int64
        trackId:
          type: string
          description: 트랙 ID
          example: sdakfjiosdwn1dnfsd
        trackName:
          type: string
          description: 트랙 제목
          example: 바다의 왕자
        artistName:
          type: string
          description: 가수이름
          example: 박명수
        imageUrl:
          type: string
          description: 이미지 URL
          example: http://dfdfjiweoif.com
    ApiResListPlaylistResponse:
      title: API 응답
      type: object
      properties:
        httpStatus:
          type: integer
          description: 응답코드
          format: int32
        data:
          type: array
          description: 응답 데이터
          items:
            $ref: "#/components/schemas/PlaylistResponse"
        message:
          type: string
          description: 응답 메세지
          example: ex.조회에 성공했습니다.
        divisionCode:
          type: string
          description: 오류 구분 코드
          example: 오류시 출력
        errors:
          type: array
          description: 상세 필드 에러
          example: 오류시 출력
          items:
            $ref: "#/components/schemas/FieldErrors"
    ApiResPlaylistTrackResponse:
      title: API 응답
      type: object
      properties:
        httpStatus:
          type: integer
          description: 응답코드
          format: int32
        data:
          $ref: "#/components/schemas/PlaylistTrackResponse"
        message:
          type: string
          description: 응답 메세지
          example: ex.조회에 성공했습니다.
        divisionCode:
          type: string
          description: 오류 구분 코드
          example: 오류시 출력
        errors:
          type: array
          description: 상세 필드 에러
          example: 오류시 출력
          items:
            $ref: "#/components/schemas/FieldErrors"
    PlaylistTrackResponse:
      type: object
      properties:
        playlist:
          $ref: "#/components/schemas/PlaylistResponse"
        track:
          type: array
          description: 플레이리스트 트랙 정보
          example:
            - playlistTrackNumber: "1"
              trackId: sdakfjiosdwn1dnfsd
              trackName: 바다의 왕자
              artistName: 박명수
              imageUrl: http://dfdfjiweoif.com
            - playlistTrackNumber: "2"
              trackId: 다른 아이디
              trackName: 다른노래
              artistName: 다른 가수
              imageUrl: 다른 링크
          items:
            $ref: "#/components/schemas/TrackResponse"
    ApiResPlaylistTrackEachResponse:
      title: API 응답
      type: object
      properties:
        httpStatus:
          type: integer
          description: 응답코드
          format: int32
        data:
          $ref: "#/components/schemas/PlaylistTrackEachResponse"
        message:
          type: string
          description: 응답 메세지
          example: ex.조회에 성공했습니다.
        divisionCode:
          type: string
          description: 오류 구분 코드
          example: 오류시 출력
        errors:
          type: array
          description: 상세 필드 에러
          example: 오류시 출력
          items:
            $ref: "#/components/schemas/FieldErrors"
    PlaylistTrackEachResponse:
      type: object
      properties:
        playlistTrackNumber:
          type: integer
          description: 플리 속 트랙 번호
          format: int64
          example: 1
        trackId:
          type: string
          description: 트랙 ID
          example: sdakfjiosdwn1dnfsd
        trackName:
          type: string
          description: 트랙 제목
          example: 바다의 왕자
        artistName:
          type: string
          description: 가수이름
          example: 박명수
        imageUrl:
          type: string
          description: 이미지 URL
          example: http://dfdfjiweoif.com
  securitySchemes:
    Authorization:
      type: http
      scheme: bearer
      bearerFormat: JWT